[FIX] Align set of known file types with runtime (#337)
UI5 runtime knows JavaScript file types controller, designtime,
fragment, support and view.

TODO: support files should not be uglified (to keep the rules source
code readable). It has to be decided whether this is realized by
configuration or whether it should be hard coded in the task / processor
or (library) type.