Migrate change 3312369 from Java tooling:
[FIX] Infra: improve recognition of main module in case of bundles

So far, the module analyzer always assumed the first module definition
to be the main definition. This failed at least for CVOM bundles but
might also fail for others.

The analysis has been improved and now accepts a module as main module
only in the following cases:
- if it is an unnamed module
- if it is a named module whose name matches the external name
- if it is a named module and there's only one module (no bundle)

If multiple module definitions match the above criteria, this is
reported as an error.

Additionally, AMD special dependencies ('require', 'exports', 'module')
are no longer reported in the dependency info.

This fixes all test, except the EVO bundle test.