Migrate change 3945349 from Java tooling:
FIX] Infra: JSModuleAnalyzer fails on modules that don't declare a name

Basically, the JavaScript implementation was less sensitive to this
issue. But if a resource could have a dependency to another resource
named "null", the JavaScript tooling would fail, too. Although this
looked like a purely theoretical scenario, adding a protective 'if' was
trivial enough to just do it.