Add included/excludedDependencies to builder
When setting the buildDependencies parameter to "true" builder
used to build all dependencies of the project. By adding two new
parameters includedDependencies and excludedDependencies,
similar to the existing included/excludedTasks parameters, one
can now choose which depenendencies to include into the build.

This is also useful for self-contained builds to include the built
depenendy into the bundle, instead of the unbuilt source read
from the file system (e.g. in case a transpile task is present).