[PoC] ResourcePool: Add option to prefer analyzing a resource's source content
Possible follow up of https://github.com/SAP/ui5-builder/pull/551 to
improve dependency analysis.

JSModuleAnalyzer has trouble detecting some dependencies in minified
code if they have been "mangled". E.g. in case jQuery is not named
jQuery anymore.

This change solves that by trying to retrieve the debug ("source")
content of a resource to analyze that instead of the minified version.
However, integrating this on the ResourcePool level is not ideal and a
more general solution to this problem should be preferred.

Therefore this PR should only act as an illustration of the discussions
and thoughts. Currently it should not be merged.