[FIX] lbt/resources/ResourceCollector: Ensure proper matching of indicator files
Indicator files like "Component.js", "manifest.json" or ".library" should only be matched
if the full basename is matched.

Before this fix, files like "MyComponent.js" were matched which resulted in the unexpected
creation of a resources.json file. The file should only be created for components, libraries
and themes.