[FEATURE] Introduce concept of task identifiers to AbstractBuilder
Task identifiers (in short taskIds) can now be used in AbstractBuilder to define the order of multiple task executions. So far the name of the task (taskName) was treated the identifier of a task. If the same task was added to the execution multiple times, an ID in form taskName--1 was generated, counting upwards from one. In the task definition there was no way of defining the sub-order of multiple executions.

Now with the introduction of task identifiers, one can define a { id: "taskA", ... } in the task definition or call addTask with both an identifier and the name of the task to call. The old logic of generating an identifier, in case only a name is defined stays in place.

The before/afterTask declarations now no longer refer to task names but to task identifiers instead. However due to task names have previously been treated as task identifiers exclusively and also the old incrementation logic for task names stayed the same, this change is compatible with any old configuration.