[INTERNAL] lib/jsdoc: do not write 'module' info for borrowed APIs
When an API is borrowed by another entity, JSDoc creates a deep clone of
the original symbol and makes it a member of the borrowing entity.

As the origin info ('module', 'resource' and 'export') is also cloned
but usually doesn't match the origin info of the borrowing entity, the
template so far decided to write out 'module' and 'export' properties
for all borrowed APIs.

This change tries to detect such clones (by remembering the symbol's
longname when the origin info is attached) and filters them out when
writing the origin info for a member.

As origin info for members currently is not evaluated anywhere, this
change should have no side effects. But it helps to clean-up the schema
for api.json files as an origin info for members has to be allowed only
in very special cases, not for all members.

Cherry picked from SAP/openui5@9d3fccd91.