[FIX] lbt/Resolver: ensure stable module order in generated bundleInfo
When a section of type 'bundleInfo' is generated as part of a bundle,
the order of the modules in each bundle info is not relevant, but should
be stable across different runs.

So far, the order of modules was dependent on the module iteration order
of the resource pool and could be random due to I/O timing.