[INTERNAL] lib/processors/jsdoc: Consistency check - throw Error
Any inconsistency of the visibility setting in the runtime
metadata and the JSDoc visibility should break the build.

Migrating from `future()` error log to `error()` log, which will cause
the build to fail at the end (if it's configured to fail on JSDoc
errors).

Cherry picked from SAP/openui5@2f3c33c53.