[INTERNAL] generateThemeDesignerResources: Allow core .theming in sources
This change respects an sap.ui.core library .theming file from the
sources instead of simply overwriting the file.
It allows the configuration to be put into the library and adjusted
there instead of having to adjust it in the build task.

The existing logic is kept to support older UI5 versions where the file
is not available in the sources.

JIRA: CPOUI5FOUNDATION-862
(cherry picked from commit dda30114461314e81da43bd739e709186744f592)