[INTERNAL] JSDoc: Fix reference to jsdoc template
jsdoc is not able to resolve the @ui5/builder dependency when there is
no @ui5/builder within node_modules available. This only became an issue
once @ui5/project switched to an optional peerDependency for
@ui5/builder so when testing the @ui5/builder itself, there was no
transitive @ui5/builder dependency available.

This also solves the issue of not testing the current version of the
template as before it came from the installed @ui5/builder version.

This also updates the package-lock.json by re-generating it.