[FIX] manifestEnhancer: Only use valid files for supportedLocales
This fixes two problems that could have occurred:

A properties file with an invalid locale was still taken into the list
of supported locales, which then caused a runtime exception in the
ResourceBundle as it validates the input.

Another problem was that properties files could have a valid name
according to BCP47, but the file won't be ever requested with that name.
This is due to the fact that the ResourceBundle does not use the
extension / private use sections of the locale when creating the request
URL.

In both cases, the properties file is now ignored and no entry for the
supportedLocales is created.