[FEATURE] Add the 'flexBundle'-flag into the manifest in the generateFlexChangesBundle task
Currently, the runtime checks the existence of the flexibility-bundle.json depending on the preloading of the bundle itself.
This preload of all component sources may not be in time for annotation changes, which are part of the preprocessing and require a more explicit hint, if the bundle is present and should be requested.
This commit adds this hint with a flag in the "sap.ui5" subsection in the manifest.